+++ lib/animate_do.dart
@@ -8,14 +8,15 @@
 /// using only native widgets. This library is inspired by Animate.css.
 library animate_do;
 
-export 'src/animations/animate_do_attention_seekers.dart';
+
+export 'src/animations/attention_seekers/index.dart';
+
 export 'src/animations/animate_do_bounces.dart';
 export 'src/animations/animate_do_elastics.dart';
 export 'src/animations/animate_do_fadeouts.dart';
 export 'src/animations/animate_do_fades.dart';
 export 'src/animations/animate_do_flips.dart';
 export 'src/animations/animate_do_slides.dart';
-export 'src/animations/animate_do_specials.dart';
 export 'src/animations/animate_do_zooms.dart';
 
 /// Types