+++ lib/src/animations/attention_seekers/flash.dart
@@ -1,0 +1,117 @@
+import 'package:flutter/material.dart';
+
+import '../../types/animate_do_mixins.dart';
+import '../../types/animate_do_types.dart';
+
+
+/// [key]: optional widget key reference
+/// [child]: mandatory, widget to animate
+/// [duration]: how much time the animation should take
+/// [delay]: delay before the animation starts
+/// [controller]: optional/mandatory, exposes the animation controller created by Animate_do
+/// [manualTrigger]: boolean that indicates if you want to trigger the animation manually with the controller
+/// [animate]: For a State controller property, if you re-render changing it from false to true, the animation will be fired immediately
+/// [onFinish]: callback that returns the direction of the animation, [AnimateDoDirection.forward] or [AnimateDoDirection.backward]
+/// [curve]: curve for the animation
+/// [infinite]: loops the animation until the widget is destroyed
+class Flash extends StatefulWidget {
+  final Widget child;
+  final Duration duration;
+  final Duration delay;
+  final bool infinite;
+  final Function(AnimationController)? controller;
+  final bool manualTrigger;
+  final bool animate;
+  final Function(AnimateDoDirection direction)? onFinish;
+  final Curve curve;
+
+  Flash(
+      {key,
+      required this.child,
+      this.duration = const Duration(milliseconds: 1000),
+      this.delay = const Duration(milliseconds: 0),
+      this.infinite = false,
+      this.controller,
+      this.manualTrigger = false,
+      this.animate = true,
+      this.onFinish,
+      this.curve = Curves.linearToEaseOut})
+      : super(key: key) {
+    if (manualTrigger == true && controller == null) {
+      throw FlutterError('If you want to use manualTrigger:true, \n\n'
+          'Then you must provide the controller property, that is a callback like:\n\n'
+          ' ( controller: AnimationController) => yourController = controller \n\n');
+    }
+  }
+
+  @override
+  FlashState createState() => FlashState();
+}
+
+/// State class, where the magic happens
+class FlashState extends State<Flash>
+    with SingleTickerProviderStateMixin, AnimateDoState {
+  late Animation<double> opacityOut1;
+  late Animation<double> opacityIn1;
+  late Animation<double> opacityOut2;
+  late Animation<double> opacityIn2;
+  @override
+  void dispose() {
+    disposed = true;
+    controller.dispose();
+    super.dispose();
+  }
+
+  @override
+  void initState() {
+    super.initState();
+
+    controller = AnimationController(duration: widget.duration, vsync: this);
+
+    opacityOut1 = Tween<double>(begin: 1, end: 0).animate(CurvedAnimation(
+        parent: controller, curve: Interval(0, 0.25, curve: widget.curve)));
+    opacityIn1 = Tween<double>(begin: 0, end: 1).animate(CurvedAnimation(
+        parent: controller, curve: Interval(0.25, 0.5, curve: widget.curve)));
+    opacityOut2 = Tween<double>(begin: 1, end: 0).animate(CurvedAnimation(
+        parent: controller, curve: Interval(0.5, 0.75, curve: widget.curve)));
+    opacityIn2 = Tween<double>(begin: 0, end: 1).animate(CurvedAnimation(
+        parent: controller, curve: Interval(0.75, 1, curve: widget.curve)));
+
+    /// Provided by the mixing [AnimateDoState] class
+    configAnimation(
+      delay: widget.delay,
+      animate: widget.animate,
+      manualTrigger: widget.manualTrigger,
+      infinite: widget.infinite,
+      onFinish: widget.onFinish,
+      controllerCallback: widget.controller,
+    );
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    /// Provided by the mixing [AnimateDoState] class
+    buildAnimation(
+      delay: widget.delay,
+      animate: widget.animate,
+      manualTrigger: widget.manualTrigger,
+      infinite: widget.infinite,
+      onFinish: widget.onFinish,
+      controllerCallback: widget.controller,
+    );
+
+    return AnimatedBuilder(
+        animation: controller,
+        builder: (BuildContext context, Widget? child) {
+          return Opacity(
+              opacity: (controller.value < 0.25)
+                  ? opacityOut1.value
+                  : (controller.value < 0.5)
+                      ? opacityIn1.value
+                      : (controller.value < 0.75)
+                          ? opacityOut2.value
+                          : opacityIn2.value,
+              child: widget.child);
+        });
+  }
+}