+++ lib/src/animations/attention_seekers/heart_beat.dart
@@ -1,0 +1,111 @@
+import 'package:flutter/material.dart';
+
+import '../../types/animate_do_mixins.dart';
+import '../../types/animate_do_types.dart';
+
+
+
+
+/// HeartBeat
+/// Class [HeartBeat]:
+/// [key]: optional widget key reference
+/// [child]: mandatory, widget to animate
+/// [duration]: how much time the animation should take
+/// [delay]: delay before the animation starts
+/// [controller]: optional/mandatory, exposes the animation controller created by Animate_do
+/// [manualTrigger]: boolean that indicates if you want to trigger the animation manually with the controller
+/// [animate]: For a State controller property, if you re-render changing it from false to true, the animation will be fired immediately
+/// [infinite]: loops the animation until the widget is destroyed
+/// [onFinish]: callback that returns the direction of the animation, [AnimateDoDirection.forward] or [AnimateDoDirection.backward]
+/// [curve]: curve for the animation
+class HeartBeat extends StatefulWidget {
+  final Widget child;
+  final Duration duration;
+  final Duration delay;
+  final Function(AnimationController)? controller;
+  final bool manualTrigger;
+  final bool animate;
+  final bool infinite;
+  final Function(AnimateDoDirection direction)? onFinish;
+  final Curve curve;
+
+  HeartBeat({
+    Key? key,
+    required this.child,
+    this.duration = const Duration(milliseconds: 2500),
+    this.delay = Duration.zero,
+    this.controller,
+    this.manualTrigger = false,
+    this.animate = true,
+    this.infinite = false,
+    this.onFinish,
+    this.curve = Curves.easeOutQuad,
+  }) : super(key: key) {
+    if (manualTrigger == true && controller == null) {
+      throw FlutterError('If you want to use manualTrigger:true, \n\n'
+          'Then you must provide the controller property, that is a callback like:\n\n'
+          ' ( controller: AnimationController) => yourController = controller \n\n');
+    }
+  }
+
+  @override
+  HeartBeatState createState() => HeartBeatState();
+}
+
+class HeartBeatState extends State<HeartBeat>
+    with SingleTickerProviderStateMixin, AnimateDoState {
+  late Animation<double> scale;
+
+  @override
+  void dispose() {
+    disposed = true;
+    controller.dispose();
+    super.dispose();
+  }
+
+  @override
+  void initState() {
+    super.initState();
+
+    controller = AnimationController(duration: widget.duration, vsync: this);
+
+    scale = TweenSequence<double>([
+      TweenSequenceItem(tween: Tween(begin: 1.0, end: 1.3), weight: 14),
+      TweenSequenceItem(tween: Tween(begin: 1.3, end: 1.0), weight: 14),
+      TweenSequenceItem(tween: Tween(begin: 1.0, end: 1.3), weight: 14),
+      TweenSequenceItem(tween: Tween(begin: 1.3, end: 1.0), weight: 28),
+      TweenSequenceItem(tween: Tween(begin: 1.0, end: 1.0), weight: 30),
+    ]).animate(CurvedAnimation(parent: controller, curve: widget.curve));
+
+    configAnimation(
+      delay: widget.delay,
+      animate: widget.animate,
+      manualTrigger: widget.manualTrigger,
+      infinite: widget.infinite,
+      onFinish: widget.onFinish,
+      controllerCallback: widget.controller,
+    );
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    buildAnimation(
+      delay: widget.delay,
+      animate: widget.animate,
+      manualTrigger: widget.manualTrigger,
+      infinite: widget.infinite,
+      onFinish: widget.onFinish,
+      controllerCallback: widget.controller,
+    );
+
+    return AnimatedBuilder(
+      animation: controller,
+      builder: (BuildContext context, Widget? child) {
+        return Transform.scale(
+          scale: scale.value,
+          child: widget.child,
+        );
+      },
+    );
+  }
+}