+++ lib/src/animations/attention_seekers/swing.dart
@@ -1,0 +1,100 @@
+import 'package:flutter/material.dart';
+
+import '../../types/animate_do_mixins.dart';
+import '../../types/animate_do_types.dart';
+
+
+
+
+// Swing 
+/// A widget that creates a swinging animation effect on its child widget.
+///
+/// [key]: optional widget key reference
+/// [child]: mandatory, widget to animate
+/// [duration]: how much time the animation should take
+/// [delay]: delay before the animation starts
+/// [controller]: optional/mandatory, exposes the animation controller created by Animate_do
+/// [manualTrigger]: boolean that indicates if you want to trigger the animation manually with the controller
+/// [animate]: For a State controller property, if you re-render changing it from false to true, the animation will be fired immediately
+/// [onFinish]: callback that returns the direction of the animation, [AnimateDoDirection.forward] or [AnimateDoDirection.backward]
+/// [curve]: curve for the animation
+/// [infinite]: loops the animation until the widget is destroyed
+class Swing extends StatefulWidget {
+  final Widget child;
+  final Duration duration;
+  final Duration delay;
+  final bool infinite;
+  final Function(AnimationController)? controller;
+  final bool manualTrigger;
+  final bool animate;
+  final Function(AnimateDoDirection direction)? onFinish;
+  final Curve curve;
+
+  Swing({
+    key,
+    required this.child,
+    this.duration = const Duration(milliseconds: 1000),
+    this.delay = const Duration(milliseconds: 0),
+    this.infinite = false,
+    this.controller,
+    this.manualTrigger = false,
+    this.animate = true,
+    this.onFinish,
+    this.curve = Curves.easeOut,
+  }) : super(key: key);
+
+  @override
+  SwingState createState() => SwingState();
+}
+
+class SwingState extends State<Swing> with SingleTickerProviderStateMixin, AnimateDoState {
+  late AnimationController controller;
+  late Animation<double> rotation;
+
+  @override
+  void initState() {
+    super.initState();
+
+    controller = AnimationController(duration: widget.duration, vsync: this);
+
+    rotation = TweenSequence<double>([
+      TweenSequenceItem(tween: Tween(begin: 0.0, end: 0.261799), weight: 20), // 15 degrees
+      TweenSequenceItem(tween: Tween(begin: 0.261799, end: -0.174533), weight: 20), // -10 degrees
+      TweenSequenceItem(tween: Tween(begin: -0.174533, end: 0.087266), weight: 20), // 5 degrees
+      TweenSequenceItem(tween: Tween(begin: 0.087266, end: -0.087266), weight: 20), // -5 degrees
+      TweenSequenceItem(tween: Tween(begin: -0.087266, end: 0.0), weight: 20), // 0 degrees
+    ]).animate(CurvedAnimation(parent: controller, curve: widget.curve));
+
+    configAnimation(
+      delay: widget.delay,
+      animate: widget.animate,
+      manualTrigger: widget.manualTrigger,
+      infinite: widget.infinite,
+      onFinish: widget.onFinish,
+      controllerCallback: widget.controller,
+    );
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    buildAnimation(
+      delay: widget.delay,
+      animate: widget.animate,
+      manualTrigger: widget.manualTrigger,
+      infinite: widget.infinite,
+      onFinish: widget.onFinish,
+      controllerCallback: widget.controller,
+    );
+
+    return AnimatedBuilder(
+      animation: controller,
+      builder: (BuildContext context, Widget? child) {
+        return Transform(
+          transform: Matrix4.identity()..rotateZ(rotation.value),
+          alignment: Alignment.topCenter,
+          child: widget.child,
+        );
+      },
+    );
+  }
+}